configure embedded Tomcat and DispatcherServlet
To start the application most easily we use an embedded Tomcat. This
means we can just run the application as a normal Java application
without any need to deploy to an external server.

We also provide the initial configuration of the DispatcherServlet that
will be used to route incomming requests to our handler methods.